Send bash xtrace output to stdout
Use BASH_XTRACEFD to send bash's "set -x" tracing output
to stdout rather than the default of stderr.  Buildbot
highlights text on stderr in red assuming that they are warnings.
It would be nice to reserve that for real warnings so we don't
get in the habit of ignoring the red text.