Builders download and build from tarball.
Change the builders to download the tarball produced
by the tarball builder instead of cloning the lustre
repository. The builders then build from the contents of
the tarball. This reduces the amount of bandwidth
required for each build.

This is phase two of the bandwidth reduction issue. Phase
one introduced the tarball builder which we leverage here.

This closes #12.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>